# CoreLocalArray class