Moved the logic for core-local array out of ConcurrentArena and into a separate class because I want to reuse it for core-local stats.

Test Plan:

setup:
- command: `TEST_TMPDIR=/dev/shm ./db_bench -benchmarks=fillrandom -threads=16 -num=1000000 --allow_concurrent_memtable_write`
- filesystem: tmpfs

before this diff:
`fillrandom   :       2.934 micros/op 340831 ops/sec;   37.7 MB/s`

after this diff:
`fillrandom   :       2.944 micros/op 339659 ops/sec;   37.6 MB/s`