@yiwu-arbug 2017-05-11
Looks like Windows build is failure. @ajkr mind send a fix? Thanks!